+++ R/readMSData2.R
@@ -18,6 +18,7 @@ readOnDiskMSData <- function(files, pdata, msLevel., verbose,
                              centroided., smoothed.) {
     .testReadMSDataInput(environment())
     stopifnot(is.logical(centroided.))
+
     ## Creating environment with Spectra objects
     assaydata <- new.env(parent = emptyenv())
     filenams <- filenums <- c()
@@ -98,7 +99,6 @@ readOnDiskMSData <- function(files, pdata, msLevel., verbose,
                    processing = paste0("Data loaded [", date(), "]"),
                    files = files,
                    smoothed = NA)
-
     ## Create 'fdata' and 'pdata' objects
     if (is.null(pdata)) {
         .pd <- data.frame(sampleNames = basename(files))
@@ -119,7 +119,7 @@ readOnDiskMSData <- function(files, pdata, msLevel., verbose,
         fdata <- new("AnnotatedDataFrame", data = fdata)
         ## Re-order the features.
         ## fdata <- fdata[ls(assaydata), ]
-    }
+    } else fdata <- new("AnnotatedDataFrame")
 
     ## expriment data slot
     if (length(.instrumentInfo) > 1) {