+++ tests/testthat/test_utils.R
@@ -182,7 +182,7 @@ test_that("mergeSpectraAndIdentificationData", {
                                                           icol = c("file", "acquisitionnum"),
                                                           acc = "accession",
                                                           desc = "description", pepseq = "pepseq")
-    expect_equal(ans1, rfd1)    
+    expect_equal(ans1, rfd1)
     ## second run
     ans2 <-
         MSnbase:::utils.mergeSpectraAndIdentificationData(rfd1, id2,
@@ -190,7 +190,7 @@ test_that("mergeSpectraAndIdentificationData", {
                                                           icol = c("file", "acquisitionnum"),
                                                           acc = "accession",
                                                           desc = "description", pepseq = "pepseq")
-    expect_equal(ans2, rfd2)    
+    expect_equal(ans2, rfd2)
 })
 
 test_that("utils.idSummary", {
@@ -516,4 +516,70 @@ test_that("windowIndices works", {
     expect_equal(windowIndices(4, 1, 5), c(3, 4, 5))
 
     expect_error(windowIndices(4, 2, 1))
+})
+
+test_that(".fix_breaks works", {
+    rtr <- c(1, 12)
+    brks <- seq(rtr[1], rtr[2], by = 4)
+    ## brks do not include rtr
+    expect_true(all(rtr[2] > brks))
+    brks_f <- .fix_breaks(brks, rtr)
+    expect_true(rtr[2] <= max(brks_f))
+
+    ## Next unit tests taken from breaks_Spectrum.
+    ## Issue #191
+    ints <- 1:4
+    brks <- seq(min(ints), max(ints), by = 1)
+    expect_equal(MSnbase:::.fix_breaks(brks, range(ints)), 1:5)
+    expect_true(any(MSnbase:::.fix_breaks(1:2, range(ints)) < 4))
+    expect_equal(MSnbase:::.fix_breaks(seq(1, 4, by = 2), range(ints)),
+                 c(1, 3, 5))
+
+    ## Test with values smaller than 1
+    rng <- c(0.1, 0.4)
+    brks <- seq(rng[1], rng[2], by = 0.04)
+    MSnbase:::.fix_breaks(brks, rng)
+})
+
+test_that(".bin_values works", {
+    set.seed(123)
+    vals <- abs(rnorm(20, mean = 40))
+    xs <- seq(1:length(vals)) + rnorm(length(vals), sd = 0.001)
+
+    res <- .bin_values(vals, xs, binSize = 1)
+    brks <- seq(0, 20, by = 1)
+    for (i in 1:(length(brks) - 1)) {
+        idx <- which(xs > brks[i] & xs < brks[i +1])
+        if (length(idx))
+            expect_equal(res$x[i], max(vals[idx]))
+        else expect_equal(res$x[i], 0)
+    }
+
+    ## Ensure that all values are within.
+    xs <- seq(1:length(vals))
+    brks <- seq(1, 20, by = 3)
+    ## brks does not contain all values.
+    expect_true(max(brks) < max(xs))
+    res <- .bin_values(vals, xs, binSize = 3, fun = sum)
+    ## The largest bin should contain all values larger than max(brks)
+    expect_equal(res$x[length(res$x)], sum(vals[xs >= max(brks)]))
+
+    ## Check exceptions
+    expect_error(.bin_values(1:3, 1:5))
+    expect_error(.bin_values(1:3, 1:5), fun = other)
+})
+
+test_that("merging and expanding features", {
+    data(hyperLOPIT2015)
+    fv <- sort(fvarLabels(hyperLOPIT2015))
+    hl <- mergeFeatureVars(hyperLOPIT2015, fcol = k, fcol2 = "SVM")
+    n <- length(fvarLabels(hl))
+    expect_identical(n, length(fv) - 2L)
+    hl2 <- expandFeatureVars(hl, "SVM", prefix = NULL)
+    fv2 <- sort(fvarLabels(hl2))
+    expect_identical(fv, fv2)
+
+    k <- grep("^svm", fvarLabels(hyperLOPIT2015), value = TRUE)
+    k2 <- grep("^svm", fvarLabels(hl2), value = TRUE)
+    expect_identical(fData(hyperLOPIT2015)[, k], fData(hl2)[, k2])
 })